pyproject.toml
[project]
name = "ultimate-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.0",
    "fastapi>=0.110.0",
    "httpx>=0.26.0",
    "pydantic>=2.6.1",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "selenium==4.17.2",
    "sqlalchemy>=2.0.27",
    "uvicorn>=0.27.1",
    "webdriver-manager==4.0.1",
]
